Added New Waiting List Component and Navigational Tabs in Header
Added New components and sub-components for pool table player waiting
list.   player waiting list collects a list of players that are waiting
to play and distributes them to the amount of tables available.  The
players are automatically placed back on the list once they've lost
their game on the table.

Added in navigational buttons in Header component as well. You can view
the default Score Sheet Component or the newly added Player Waiting list
component.