Changed Match Component view, also included a TODO
Modified the Games section of the  Match Component.  Instead of
cluttering the ladnscape with mulitple games displayed at once, only
displaying one/current game at a time.  Users will be able to
view/modify previous games of the match if the label is clicked.
TODO: Must work on correctly reflecting the '8 on Break' and 'Break and
Run' values correctly.  Currently the passed game data isn't properly
updating the checkbox selected